pyproject.toml
[project]
name = "budgetmoney"
version = "0.1.0"
description = "A Rocket Money transaction analyzer tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-core>=2.24.0",
    "google-api-python-client>=2.158.0",
    "google-auth>=2.37.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "ipykernel>=6.29.5",
    "jupyterlab>=4.3.4",
    "langchain>=0.3.14",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.2.2",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.6",
    "streamlit>=1.41.1",
    "transformers>=4.47.1",
    "typer>=0.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
bmoney = "budgetmoney.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["budgetmoney"]
